Add `PhoenixTest.submit/1`
What changed?
=============

We add a new `submit/2` function to handle form submissions _without_
explicitly clicking a button. It is intended to emulate `<Enter>`
behavior.

This is intended to be used with active forms. If there is no active
form, we'll raise. Thus, this is meant to be used with `fill_in`,
`select`, `choose`, etc. to replace what people do with `submit_form`
today.

We also make it so that `submit/1` includes the first button's `name`
and `value` (if a button is present) as part of the form data. That
emulates