Fix fill_in, etc. polluting across different forms
What changed?
============

There are scenarios where we might use two `fill_in`s back-to-back to
fill different forms (e.g. a select that populates another select's
options that live in different forms).

Currently, any `fill_in` after the first one will just add data to the
"active_form" data. But that results in errors because the second form
doesn't have the first form's fields.

In order to handle that, every field action now checks to see if the
form we're acting on is the same as the previous one.

As part of this work, we refactor field actions (`fill_in`, `select`,
etc.) to not require active form and to share a `fill_in_field_data`
helper where we encapsulate the new form check.