Fix: handle nested forms with field functions
What changed?
=============

We recently refactored our field functions (`fill_in`, `select`, etc.)
to not rely on ActiveForm where it wasn't needed. We replaced the
`ActiveForm.add_form_data` with `Map.merge` in several places.

But those aren't the same. ActiveForm was performing a deep merge (which
helps us deal with nested forms).

We add a test to ensure we do that and fix the bugs by using DeppMerge
whenever we merge form data.