assert_path/refute_path partially handle Live Navigation
What changed?
=============

We improve our `assert_path` and `refute_path` implementation to
partially handle live navigation. For now, we handle `<.link navigate>`
options but we do not support `<.link patch>`

Patching is done at the LiveView process level, and we need to figure
out how to best handle that.

For now, we release this fix to improve the experience, but we include a
warning in the docs that the helpers do not support patching.