Handle multiple checkboxes in inside label
What changed?
============

Our `Query.find_by_label` function handles implicit and explicit labels
(implicit being those where the input is associated to the label by
nesting instead of for<->id). But it does not handle the case when there
are multiple inputs under a single label.

When would that happen?
----------------------

Checkboxes typically have two inputs nested: the normal one and a hidden
one that is submitted in case the checkbox is NOT checked.

This commit updates the our search to ignore hidden inputs when
searching for implicit labels.