Do not pass value to `Field.find_input!/3`
What changed?
============

Our `Field.find_input!/3` behaves differently from the rest of the
functions in the module in that it allows for passing a value that
overrides what we find in the input itself.

We change that for 2 reasons:

1.That's confusing behavior since we're no longer just "finding" the
  input. We're setting a value to it.
2.The override doesn't work if we set the input's value to `nil`. That's
  something we may want to do if testing an update to `nil` (e.g. to see
  errors `fill_in("Label", with: nil)`

So, for now we keep `find_input!/2` and `Map.put/3` the value on top.