Fix `Form.find_by_descendant!` to find with field
What changed?
============

Our `Form.find_by_descendant!` worked for buttons and field so long as
the fields had an id. If the field didn't have an id, the form code
expected an element with `selector` and `text` (fields have no text). We
update the code to use text only if available.