Add .git-blame-ignore-revs
We add .git-blame-ignore-revs so we can mark some commits as ignored by
git blame. That's helpful since we added Styler and mix formatted
everything. A nice bonus is that GitHub also picks up those files.

Thanks to https://www.stefanjudis.com/today-i-learned/how-to-exclude-commits-from-git-blame/
for the tips on how to do that.