Refactor: Form doesn't depend on Html.Form
What changed?
=============

`Html.Form` was an original pass at parsing forms. Since then we've moved
more logic into `Form`, but we still depended on `Html.Form` to
determine the "operative" method. We now move that logic to `Form` and
clean up the tests to prepare us to (hopefully) remove the need for
`Html.Form`.