Refactor: Do not call deprecated functions internally
What changed?
=============

We update our `Live` and `Static` implementations so they do not call
deprecated functions (`fill_form` and `submit_form`) internally.

There are still many deprecation warnings due to tests we have testing
that logic. But we'll remove those tests when we bump the version.

Finally, in order to make this work, we move the Live and Static logic
directly into the modules rather than defining that in the `defimpl`
portions. That allows us to call the functionality directly when we
want, and the protocol implementations just delegate the function calls.