Move README docs to PhoenixTest (and augment)
What changed?
============

The library's surface area is too small for a separate README as an
"introduction" in ExDoc's "Pages" section.

More often than not, we want to find the setup, installation, etc. AND
the exact function definitions.

For that reason, we move the README docs into the `PhoenixTest` module.
And as we do so, we take the opportunity to update them with a couple of
sections on filling out forms.