Add: handle redirects on `phx-change` forms
What changed?
============

It's possible for forms to `push_navigate` (or do some other redirect)
on a `phx-change` event. So far, our Live implementation doesn't handle
redirects there.

This commit updates that code to be more in line with what happens with
`phx-submit` -- it calls `maybe_redirect` to handle potential redirects.