Remove Html.Form (and validation)
What changed?
============

Now that we no longer expose `fill_form` and `submit_form`, we don't
need to validate that the fields passed in a map correspond to the HTML
form they're targeting.

By using `fill_in`, `select`, etc., we ensure the field exists when we
first try to fill it in. So, we can finally remove the `Html.Form`
module.