Add: `click_button` can submit forms when not nested
What changed?
=============

When a button has the `form` attribute (pointing to its form owner's
ID), we should be able to submit the form by clicking on that button.

In fact, even if the button is nested within another form, the `form`
attribute takes precedence in which form it'll submit.

This change makes it so that we search for that "form owner" when
clicking submit instead of just assuming the button will be nested
within an ancestor form.