✨ Add `PhoenixTest.unwrap/2` escape hatch
What changed?
=============

We add a new `PhoenixTest.unwrap/2` function that provides an escape
hatch when people need to reach out for the `view` or (probably less
likely) for the `conn`.

It's not hard to imagine a simple scenario where this is needed --
LiveView has `render_hook` functionality that allows people to test
`phx-hook`.

But we can't really reproduce that with PhoenixTest.

It's not something a user always "clicks" or "visits". Sometimes, you
just happen to move your browser to that point and you have a `phx-hook`
that triggers when the element is mounted.

So, for now we provide the escape hatch.

But in order to keep things moving nicely, we handle any potential
redirects after the "unwrapped" action is performed. That allows us to
continue the pipeline nicely.