♻️  move current_path tests into Live/Static tests
What changed?
=============

We split the tests for `assert_path`/`refute_path`.

We keep the tests for assertions and refutations given a set
`session.current_path`. We then move the Live and Static tests making
sure that they're updating the `session.current_path` as expected.

In this way, our tests abide by a contract. We lose the integration
aspect of them, but that seems okay given the contract is the
`session.current_path`.