`refute_has/2` doesn't handle `:at` option without `text`
What changed?
=============

The `refute_has/2` helper seems to handle `:at` option (for position)
well if we pass `:text` as an option. But if we don't pass `text`, the
internal `finder_fun` seems to ignore the options.

This commit updates that so we pass the options down.

NOTE: we have to update the `Query.find` function heads so the compiler
doesn't get confused between `find/3` and `find/4` with defaults.
Ideally, we should combine the two by using the `text:` option in
`Query.find` and not outside of it.