Use `[id=<id>]` when querying by ID
Resolves https://github.com/germsvel/phoenix_test/issues/97

What changed?
=============

In b79e8a3, we started work to query by id using the `[id=<id>]` format.
Floki cannot find a query by selector using `#id` if the `id` in
question has a question mark.

Turns out we didn't cover all the cases in that commit, so we handle the
rest of the cases when we query by id.