Find buttons by role
What changed?
============

We update `click_button/2` to attempt to find all possible buttons by
role. That includes `<button>` tags but also things like
`input[type="button"]`, `input[type="submit"]`, etc.

In order to support that, we rename our `Selectors` into `Locators` and
add logic there. The idea is for this module to remain private for now.
We may want to improve the data structures created there.

But for now, `Locators.button` returns a `{:button, data}` tuple where
we store the roles and text. And then `Locators.role_selectors` gets all
the selectors needed. Then, we can call a new `Query.find_by_role!`
function that takes in a "locator" and gets a list of roles to find one.

Finally, we update how `Query.find_one_of` works. Previously, it was
incorrectly only finding results for one of the selectors passed (which
sometimes could be many results). Now, we're filtering all the results
and returning them in a `{:found_many, results}` tuple (when there are
several). That allows `find_one_of!` and `find_by_role!` to raise when
there are too many results found.