Handle radio buttons with phx-click
What changed?
============

We add the ability for radio buttons (i.e. `input type="radio"`) to live
outside of forms in LiveViews.

Typically, people would want to add a `phx-click` to the radio button
and rely on the input's `value` when it comes through the params.

Thus, we handle that specific case. We follow the pattern established in
`click_button` to check:

1. If the input has a phx-click
2. If the input belongs to a form, or
3. Raise an error informing user that we expect the input to have one or
   the other

NOTE: we do not yet handle `phx-value-*` being passed. We might have to,
but it seems that in radio buttons, that's unlikely to be the need. By
default, Phoenix sends the input's `value`, so that's what we should
rely on.

For our tests, we use the exact radio HTML found in MDN's documentation:
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/radio