Allow select to work with phx-click (#114)
What changed?
============

We update the `select` helper to work with `phx-click` outside of forms.

Note that the `phx-click` has to be on the option being selected, not on
the `select` tag itself.

For multi-selects, we trigger a `phx-click` for each option selected.
That emulates how Phoenix receives events from the browser with
multi-select.

Restructuring
-------------

As part of this work, we extract a `Select` module to keep all the
select logic. Now `Field` could probably be renamed `Input`, since it's
dealing with inputs (text, radio, checkbox).

Finally, we move the `Field.to_form_data` function into a `FormData`
module. We originally though to make it a protocol, but it seems
overkill. For now, pattern matching just like we had in `Field` will do.