Check/uncheck work with phx-click outside forms
What changed?
=============

We update `check` and `uncheck` to work with `phx-click` attributes
outside of forms.

The `check` case is straightforward. The `uncheck` is more complicated.

For the `uncheck` case, we first look for an checkbox that is visible.
That's contrary to what we were doing with form checkboxes -- where we
need a hidden input -- because a single checkbox can have a `phx-click`
for checking and unchecking. That removes the need for the hidden input
(like forms need).

So, if we find the checkbox and it has a `phx-click` attached, we
trigger that. If not, we check if the field belongs to a form. If so, we
find the hidden "uncheckbox" to get the value.

Then, we run into the problem that Phoenix's `phx-click` behavior on a
checkbox outside of a form makes sense but `render_click` behaves
differently:

- If a checkbox is getting checked, the `value` is included in the
  params
- If a checkbox is getting un-checked, the `value` is not included in
  the params (i.e. we get an empty map)

But `view |> element |> render_click` always includes the `value` of the
element in the payload. To bypass that, we have to avoid using the
`element` helper and directly send the event to `render_click/3` along
with an empty map for a payload.