Adds `assert_has/3` `:timeout` option
What changed?
=============

We add a `:timeout` option to `assert_has/3` that allows us to wait for
certain types of things:

1. Async LiveView processing (via `assign_async` or `start_async`). This
   is similar to what `render_async` does, but we handle it
   transparently.

2. Async redirects.

The second point has been something very difficult to handle
transparently. It's simple enough to handle from LiveViewTest via
`assert_redirect`. But we don't want to add a specific
`wait_for_redirect` or something like that since PhoenixTest handles
redirects automatically.

The one case we haven't been able to handle is when a LiveView performs
a `push_navigate` as a result of a message (i.e. `handle_info/3`) or a
result of an async operation (i.e. `handle_async/3`).

We can now handle that.

To handle both cases transparently, we create a new "watcher" process
that watches the LiveView process and any async tasks it may have
spawned. We copy (and rely) on some internal LiveViewTest logic to fetch
the async pids, so that's something we'll have to keep in line with
LiveViewTest.

We also make a few decisions about the behavior of timeout:

- Even when we pass a timeout, we first attempt an assertion on the
  current state of the page. (we rescue if it fails). That's helpful so
  that people who accidentally pass a timeout don't have to pay a
  penalty for the timeout. The one downside is that it'll cause failures
  to wait for the timeout to end. By default, that's 100ms (or ExUnit's
  default configuration)

- If someone passes timeout `0`, we honor it, and do not perform any
  async checks.

- Ignore timeout for static driver. If we have a static driver, we
  simply ignore the timeout. There's no waiting we can do there.

- Even with our "watcher" process in place, we have a potential for a
  race condition. If an async operation finishes and we notify the test
  to check the assertion again -- but the LiveView redirects as a result
  of the async operation faster than we can check, that means our test
  process would make a GenServer call (to render the LiveView) on a dead
  LiveView. To help with that, we delay the notification of async
  processes finishing just slightly. It's not a foolproof way of
  handling this. But it seems to work pretty well if the redirect
  happens as a result of the async operation.