Adds `assert_has/3` | `refute_has/3` `:timeout` option
What changed?
=============

We add a `:timeout` option to `assert_has/3` and `refute_has/3` that
allows us to wait for certain types of things in LiveViews:

1. Async LiveView processing (via `assign_async` or `start_async`). This
   is similar to what `render_async` does, but we handle it
   transparently.

2. Async redirects -- when your LiveView performs a redirect after some
   async operation in a `handle_async` or `handle_info` callback.

The second point has been something very difficult to handle
transparently. It's simple enough to handle from LiveViewTest via
`assert_redirect`. But we don't want to add a specific
`wait_for_redirect` or something like that since PhoenixTest so far
follows redirects automatically.

The one case we haven't been able to handle (until now) is when a
LiveView performs a `push_navigate` (or full redirect) as a result of a
message (i.e. `handle_info/3`) or a result of an async operation (i.e.
`handle_async/3`).

We can now handle that (with some dark magic).

Implementation notes
--------------------

To handle cases (1) and (2) transparently (from the user's perspective),
we create a new "watcher" process that watches the LiveView process and
any async tasks it may have spawned. We copy (and rely) on some internal
LiveViewTest logic to fetch the async pids, so that's something we'll
have to keep in line with LiveViewTest.

We also make a few decisions about the behavior of timeout:

- We ignore timeout for the Static driver. If we have a static driver,
  we simply use `Assertions.assert_has/3`. There's no waiting we can do
  there.

- If someone passes timeout `0`, we honor it, and do not perform any
  async checks.

- Even with our "watcher" process in place, we have a potential for a
  race condition. If an async operation finishes, and the watcher
  notifies the test to check the assertion again -- but the LiveView
  redirects as a result of the async operation faster than we can check
  -- our test process could make a GenServer call (to render the
  LiveView) on a dead LiveView. To help with that, we delay the
  notification of async processes finishing every so slightly (grep for
  `@live_view_buffer`). It's not a foolproof way of handling this, but
  it seems to work pretty well so far.

Add docs