Add choose/3 with input selector
What changed?
=============

Much like `click_button` and `click_link` have the option to pass a CSS
selector to target an element, we create a new `choose/3` that not only
allows the user to target via label but also narrow down the scope to
include an input selector.

This is helpful in cases where you have the same label for several
fields (think "Yes"/"No") questions where each label correctly points
the `for` attribute to a different input's `id`.