Add `fill_in/4` to specify CSS selector
What changed?
============

There are instances where an input's labels are the same (even when
their `for` attributes point to different inputs).

In order to support those scenarios, we create a new `fill_in/4` that
allows the user to target an input with that CSS selector (but note that
it needs to be correctly associated to the label).

Side note: while doing this, we noticed we didn't have an independent
textarea test. So we add one in LiveTest.