♻️  Query.find_by_label takes multiple CSS selectors
What changed?
=============

In preparation to better support locators' roles, and in order to
support `fill_in`, `check`, etc. being able to pass a CSS selector in
addition to the label, we update `Query.find_by_label/3` (and the `!`
version) to take in a list of CSS selectors.

That means we also lift a search we were previously doing deep in
`Query` (for `find_one_of`) and move the selectors to the appropriate
place. It actually leaves things better. Prior to this change, we were
searching for inputs (not hidden), select, or textareas in that deep
place -- and we did that because it had to handle all potential labeled
form controls. Now, that responsibility lies in the calling site. So,
places like `find_input!` or `find_checkbox!` can pass the simpler and
correct version of the selectors.

As we did this, we also had to update the error messages (since we can
now have multiple selectors being passed). So we update the error
messages and the tests. Something worth noting is that we update the
test assertions to only include the part of the message that is
important to the test in question in a regex (`~r//`) in the
`assert_raise`. That makes our test less likely to break due to error
message changes, so long as the spirit of the message remains true.