Adds `check/3` and `uncheck/3` to allow for selectors
What changed?
=============

We add `check/3` and `uncheck/3` to allow for targeting checkboxes with
CSS selectors for cases when there are multiple labels with the same
text on the same form.

Note: we also fix some of the existing documentation to add clarity on
which selector is being passed in in `choose/3` and `fill_in/4`