Fix: Don't error on input without name attribute (#123)
If the user is trying to target a specific input, say with `fill_in`,
then we should raise a nicer error if the input doesn't have a 
name. This commit _does not_ handle that. 

But for the cases where a user just has a form, and one of the
inputs doesn't have a `name`, then we shouldn't raise an 
error when we parse the form for "default" data. That seems
like phoenix test is stretching a little and forcing people to 
have the `name` defined.