Raise nice error if field is missing `name` attribute
What changed?
============

Fields (inputs and selects) need a `name` attribute for us to submit
forms. Currently, the tests fail if we're targeting a field without a
`name`. But it fails ugly -- when we try to encode the data in the form.

This commit adds a nicer error. It validates that the fields we're
turning into form data (via `FormData.to_form_data!`) have a `name`
attribute.

Note: we also backfill a few tests on the `Field` module.