♻️  Remove unused Query code and outdated docs
We remove the unused `Query` functions (based on our previous
refactoring). And we take the chance to remove the function docs that
are sometimes right and sometimes wrong. They're outdated and generally
not how we want to keep track of documentation in this file for now.

To document inputs and outputs, it'd be better to have some typespecs
and dialyzer -- something we may want to add in the future.