🛠️ Select options are always exact
What changed?
=============

PR #118 (commit dc7ba01) introduced the ability to pass `exact: true |
false` to `select`.

That need arose because we were doing substring matches in select
options. So, if a form had the following two options:

```html
<option value="orc">Orc</option>
<option value="other_orc">Other Orc</option>
```

there was no way to distinguish between them!

To provide backwards compatibility, I suggested we include the ability
to pass the `exact: false` option. But I now think that was a bad
suggestion.

It seems to me that options should always be exact, and the previous
implementation was basically a bug.

Why not leave the `:exact` option just in case?
----------------------------------------------

Other work is introducing the ability to target labels with `exact: true
| false`. The `select` function is the only one where we need to
distinguish between an exact label and exact option. So, we were going
to change the option to `:exact_option`. But really, options should
_always_ be exact.

Unlike labels, `<option>` elements tend to have simple text inside them
-- and as seen by the issue in PR #118, it makes little sense for the
option to be inexact.

Labels, on the other hand, can have more complex nested HTML inside them
(e.g. some `<span>*</span>` for required attrs). Thus, it's nice to be
able to target the label inexactly (with substring matches).

Thankfully, we haven't released this work yet, so we can change it
freely without fear of someone already relying on `exact: false`. Of
course, since we're going to make options always exact, there's a
possibility of a breaking change. We'll note that clearly in the release
notes.