Adds `select/4` `exact_option` option
What changed?
============

We add an `exact_option: true | false` option to `select/4` to allow for
the option's text to be matched exactly or inexactly.

The latest release made it into _always_ exact, but people sometimes
have complex options, and they'd like to target them inexactly. So, we
allow for that.