Encapsulate FormData data structure
What changed?
============

We encapsulate what it means to be a "form data" in the FormData module.

That prevents the rest of the codebase knowing what type of data
structure keeps track of the form data (which will be helpful when we
change it), and it makes it so we have a consistent data type across
Form and ActiveForm. Both must use FormData since it has to be combined
in the end.

Now, all form data knowledge lies in FormData. So field, select, button,
and even form delegate the knowledge of form data to `FormData`