Move name validation out of FormData parsing
What changed?
============

Now that we have a `FormData` module that's transforming a field into
its form data, it's clear that the name validation doesn't belong there.

Thus, we move it to the Static and Live modules when we're filling a
form field. The core logic lives in `Field.validate_name!` which is
called from Static and Live.