Extract payload operations into FormPayload
What changed?
============

Now that we've separated the `FormData` into its own module, we can see
clearly that the payload functionality in `Form` is not really about the
form. Instead, it's transforming `FormData` into payload data -- data
that is ready to be sent to `dispatch`.

So, we extract a new `FormPayload` module that makes that
transformation.

We also move `FormPayload` functionality out of `FormTest` and rely on
`FormData`.