Deduplicate data passed to FormPayload (#137)
What changed?
=============

Currently, `FormData` keeps the data passed as a list of tuples. That's
not an optimal strategy since it doesn't deal with duplicates well.

That being said, there are two competing behaviors that we need to
handle:

- One the one hand, we want to keep duplicates of list-like `name`s
  (e.g. `{"contact[]", "email"}`.
- On the other hand, we don't really want to keep duplicates of the same
  operation done multiple times (e.g. `[{"email", "frodo@example.com"},
  {"email", "aragorn@example.com"}]`).

In the latter case, we can imagine a user doing a `fill_in` twice, where
the email field changed twice. In the first case, we might have
something like a checkbox where the user can check multiple values.

Like we said, a list of tuples doesn't seem to be the optimal data
structure for `FormData`. But while we have it, we handle deduplication
of keys in `FormPayload.new/1`.

But we have to consider one more case. If a user calls:

```elixir
session
|> fill_in("Email", with: "frodo@example.com")
|> fill_in("Email", with: "aragorn@example.com")
|> fill_in("Email", with: "frodo@example.com")
```

The final record should show `frodo@example.com` as the `email`. If we
do a simple `Enum.uniq/1`, we dedup `frood@example.com`, leaving
`aragorn@example.com` as the last operation -- which is not what we
want. Thus, our current implementation has to reverse the list of data,
unique it, and then reverse it again.