Fix: do not recycle fresh conn
What changed?
=============

In 7ad8cb1, we recycled connection headers. To make recycling to calls
to `PhoenixTest.visit` simple, we recycled them in the top level
`visit/2` function before the `get/2` request. This has some unintended
side-effects when you have a fresh connection. So, we don't want to do
it then.