Refactoring: make supporting app more realistic
What changed?
============

The Phoenix app that lives in `test/support/` that we use to test
PhoenixTest itself has grown over time. At first, we tried to keep it
simple by only using what we needed.

That's still partially true (e.g. we don't have a full
`PhoenixTest.Web` module that we `use` from controllers, live views,
etc). But this commit tries to make it a lot more realistic to gain
confidence that the errors we get are those users of the library will
get.

Some noteworthy points:

- We make the test Endpoint more real. That showed the need to define
  error pages (and clarify  our tests around that), and to properly
  create a session in the plug pipeline.

- We move supporting modules inside `WebApp`. I considered only making
  it just `Web`, but I liked the clarity that it's web-app logic -- not
  some web-supporting code that PhoenixTest exports. I also didn't like
  `PhoenixTestWeb` since that's a different namespace and no need to
  pollute that.

- We split the root layout out of `PageView` into a `LayoutView`.

- We add an app layout in addition to the root layout. Root is always
  rendered, which includes the title. So we update tests to account for
  that.That also showed that we didn't need a "no-root" flow to test. We
  used that to test title not being rendered, but in reality, the title
  is part of the root layout, so the default will always be shown.

- The layouts now render flashes. We add some smoke tests in
  `Live.visit/2` and `Static.visit/2` to ensure we're rendering the
  flash messages.