Document flash presence in tests
We document flash presence in several tests when we redirect or navigate.

We do not include two places where the flash doesn't show up. There's an
open issue to handle flash when a LiveView performs a regular redirect.
We need to investigate that. Until then, this starts the process of
ensuring we have flash coverage when we redirect or navigate.