Use ConnHandler.visit on Live.maybe_redirect
What changed?
=============

Whenever we have a `Live.maybe_redirect` call with `{:ok, conn}`
(meaning we've followed a redirect, and it's a new static connection
instead of another LiveView), we now use `ConnHandler.visit/1`.

It's a modified version of `visit/2` that handles redirects. We do that:

(a) to keep visit logic in a central place, and
(b) to handle other potential redirects -- even though the `{:ok, conn}`
Live case already followed 1 redirect, there can be others.