Ping LiveView instead of using "buffer"
What changed?
=============

In our `timeout` implementation, we had to have a `@live_view_buffer`
that we used to delay the notification to the test process that async
processes had completed.

We did that because there was a potential race condition where an async
process completed, but the resulting operation was a LiveView redirect.

For example, that would happen like this:

```elixir
def handle_async(:some_async_task, {:ok, result}, socket) do
  # some logic
  {:noreply, redirect(socket, to: "some/path")}
end
```

If the logic in the `handle_async/3` callback was just the right length,
then our test would try to render the page (as a result of the
`:async_process_completed`) before we could notify the test that the
LiveView redirected. But LiveViewTest's `render_*` actions have a ping
to LiveView to make sure it's still alive -- and that would fail because
the LiveView is no longer alive!

That's why we had a buffer to delay the notification.

Naturally, the problem still existed -- it was just a matter of code
executing in the correct order within the right time. So, the solution
wasn't great.

After finding the error in the wild, we were able to reproduce it by
changing how our `AsyncPageLive` sleeps inside the `handle_async/3`
callback.

The fix
--------

This commit uses LiveViewTest's own method of checking if the LiveView
is alive before executing a render on it -- that is, we use the
`LiveView.Channel.ping/1` check to make sure the LiveView is alive
(which is a GenServer.call with `:infinity` timeout).

Now, there's still a potential for a race condition there, but it's far
smaller. And since we're doing the check, we can actually catch the
`:exit` in case the LiveView already dies. At that point, we could
handle the shutdown/redirect, but instead, we just call our retry
function that should receive the notification that LiveView redirected.

Finally, we change the order of messages pattern matched out of the
mailbox in the `LiveViewTimeout` module to handle
`:live_view_redirected` messages first -- just on the off chance that we
have a race condition there. But that's not the core issue here.